refactor: implement ContextWindowBuilder for improved message context handling in agent
